Added Clear Button to search bar. Moved functions to directory.
