+++ data/arrays.js
@@ -10,7 +10,6 @@ export const reposArr = [
       html: true,
       css: true
     }
-    // type: ['js', 'html', 'css']
   },
   {
     id: 2,
@@ -23,7 +22,42 @@ export const reposArr = [
       html: true,
       css: true
     }
-    // type: ['html', 'css']
+  },
+  {
+    id: 3,
+    name: `Jane's Addictions`,
+    pinned: true,
+    favorite: true,
+    description: 'Vices and Such',
+    type: {
+      js: false,
+      html: false,
+      css: true
+    }
+  },
+  {
+    id: 4,
+    name: 'Alphabet Soup',
+    pinned: false,
+    favorite: true,
+    description: 'Do wafting zephyrs quickly vex Jumbo?',
+    type: {
+      js: false,
+      html: true,
+      css: true
+    }
+  },
+  {
+    id: 5,
+    name: 'Palindromes',
+    pinned: false,
+    favorite: false,
+    description: 'No, Sir, Away! A Papaya War is On!',
+    type: {
+      js: false,
+      html: true,
+      css: false
+    }
   }
 ];
